Upgrade dependencies for rescript and refactor some code accordingly
Upgrades to ReScript 11 beta. This version is already quite stable and is an
important preparation to be able to migrate to the full v11 release later on.

Adds rescript-core as the future standard library.

Removes bs-json and its decoding logic; we use the new ReScript v11 variant
representation to directly map to our color variants instead, so might as well
just do type casts instead. This may allow runtime errors if the backend
changes its interface, but other than that the rest of the app is still sound.

Upgrades to react@18 and its latest rescript JSX version (JSX@4).